BacteriaCounterV1/requirements.txt
streamlit
opencv-python
ultralytics
numpy
Pillow